+++ chat-bot-api/recommenders/database_user_item_filtering_recommender.py
@@ -254,7 +254,7 @@ class DatabaseUserItemFilteringRecommender:
         text_query_limit               : int   = 2000,
         not_seen                       : bool  = True,
         k_sim_users                    : int   = 5,
-        random_selection_items_by_user : int   = 0.5,
+        random_selection_items_by_user : float = 0.5,
         max_items_by_user              : int   = 5,
         min_rating_by_user             : float = 3.5
     ):
